backend/requirements.txt
# 核心框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 数据库
sqlalchemy==2.0.23
mysqlclient==2.2.0
redis==5.0.1
alembic==1.13.0

# AI和机器学习
openai==1.3.7
transformers==4.36.2
torch==2.1.1
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2

# 语音处理
speechrecognition==3.10.0
pyaudio==0.2.11
librosa==0.10.1

# 情绪分析
textblob==0.17.1
vaderSentiment==3.3.2

# 安全和认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 工具库
python-dotenv==1.0.0
requests==2.31.0
httpx==0.25.2
loguru==0.7.2
aiohttp==3.9.1
websockets==12.0
cryptography==41.0.7

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1
